:bug: Fixes for Receivers

+++ types/MouseEvent.go
@@ -2,7 +2,7 @@ package types
 
 type MouseEvent struct {
 	Type   MouseEventType   `json:"type"`
-	DX     float64          `json:"dx"`
-	DY     float64          `json:"dy"`
+	DX     int              `json:"dx"`
+	DY     int              `json:"dy"`
 	Button MouseEventButton `json:"button"`
 }
